Risolto problema di valutazione tempo nella directory.

+++ Progetto_PDS/client.cpp
@@ -377,7 +377,7 @@ void send_file(boost::asio::io_service& io_service, boost::asio::basic_stream_so
 	int length, count = 0;
 	boost::posix_time::ptime start, end; //Utile per valuare il tempo di invio di un pacchetto verso il server
 	long int dif = 0, sec = 0;  //Utile per valuare il tempo di invio di un pacchetto verso il server
-	int calcola_tempo = 1; //è utile per non valutare il tempo troppe volte, ma solo ogni 50 pacchetti inviati.
+	long calcola_tempo = 1; //è utile per non valutare il tempo troppe volte, ma solo ogni 50 pacchetti inviati.
 	std::string response, send; //Risposta del server e Buffer utile all'invio di un pacchetto
 	char buf_to_send[BUFLEN];  //char utili per caricare il pacchetto di lunghezza BUFLEN da inviare al server.
 	char buf_recive[PROTOCOL_PACKET]; //Buffer utile alla ricezione di un pacchetto. La funzione da me utilizzata prevede di ricevere un char di lunghezza fissa.
@@ -464,10 +464,10 @@ void send_file(boost::asio::io_service& io_service, boost::asio::basic_stream_so
 				calcola_tempo++;
 
 				
-					event1.SetPayload(dim_send);
-					wxQueueEvent(progBar, event1.Clone());
-					event2.SetPayload(sec);
-					wxQueueEvent(progBar, event2.Clone());
+				event1.SetPayload(dim_send);
+				wxQueueEvent(progBar, event1.Clone());
+				event2.SetPayload(sec);
+				wxQueueEvent(progBar, event2.Clone());
 				
 
 				wxMutexGuiEnter();
